% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_corr.R
\name{gen_corr}
\alias{gen_corr}
\title{gen_corr}
\usage{
gen_corr(
  D = 50,
  clusters = c("none", "non-overlapping", "overlapping"),
  num_clusters = NULL,
  num_layers = NULL
)
}
\arguments{
\item{D}{The number of variables, generates an D by D correlation matrix.}

\item{clusters}{a character string specifying the type of cluster structure.
Available options are "none", for a correlation matrix with no clusters,
"non-overlapping" for a correlation matrix with one layer of clusters, and
"overlapping" for a correlation matrix with up to 4 layers and a set number
of clusters per layer.}

\item{num_clusters}{if clusters = "non-overlapping" or clusters = "none" then
num_clusters is an integer value specifying the number of clusters. If clusters =
"overlapping" then num_clusters must be a vector, arranged in descending order, of
length equal to num_layers specifying the number of clusters per layer.}

\item{num_layers}{an positive integer value between 1 and 4, specifying the number
of cluster layers. Only needed if using "overlapping" clusters.}
}
\value{
this function returns a D by D correlation matrix.
}
\description{
This function allows users to easily generate ad hoc correlation
matrices with a set number of clusters and up to 4 layers.
}
\examples{
\dontrun{
library(ggcorrplot)

### This generates a 50 by 50 correlation matrix with no clusters.
gen_corr(D = 50, clusters = "none) \%>\%
         ggcorrplot(title = "Overlapping clusters")

### This generates a 50 by 50 correlation matrix with 5 non-overlapping clusters.
gen_corr(D = 50, clusters = "non-overlapping) \%>\%
         ggcorrplot(title = "Overlapping clusters")

### This generates a 60 by 60 correlation matrix consisting
### of 4 layers with 10, 5, 3 and 2 clusters respectively.
gen_corr(D = 60,
         clusters = "overlapping",
         num_layers = 4,
         num_clusters = c(10,5,3,2)) \%>\%
                  ggcorrplot(title = "Overlapping clusters")

}
}
