% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_garch.R
\name{sim_garch}
\alias{sim_garch}
\title{sim_garch}
\usage{
sim_garch(innovations, model = list(), simple = TRUE)
}
\arguments{
\item{innovations}{a vector containing the random numbers/ the innovations of the
ARIMA + GARCH process.}

\item{model}{a list containing various ARMA + APARCH parameters, allowing once to specify the
time series properties of the simulated returns. Note that parameter combinations resulting in
non-stationary of the mean or variance will produce NAN's. The default values are set as
 list(omega = 5e-04, alpha = 0, gamma = NULL, beta = 0, mu = 0, ar = NULL, ma = NULL, delta = 2).
 See the "model" parameter under _fGarch::garchSpec()_ for more details regarding the parameters.}

\item{simple}{a logical parameter indicating if the output should be a simple vector containing just the
resulting ARIMA + GARCH series, or if FALSE a three column dataframe containing z - the innovations, h - the
 conditional variance and y - ARMA + APARCH series.}
}
\value{
if simple = TRUE a vector of the resulting ARIMA + GARCH series, else if simple = FALSE a
three column dataframe containing z - the innovations, h - the conditional variance and y - ARIMA +
GARCH series. Note the length of the resulting series will be less that that of the innovations as
 ARIMA + GARCH models require up to 5 innovations (depending on the max order) to produce the first result.
}
\description{
This function takes a vector of random numbers, refered to as innovations, and
induces mean and variance persistence by inserting them into an ARMA + APARCH model.
}
\note{
It is suggested that the randomly distributed numbers be mean zero and standard
deviation one, as those attributes are better set in the model argument.
}
\examples{

\dontrun{

library(tidyverse)

### creating series of 501 innovations
inno <-  rnorm(501)

### This produces a ARIMA + GARCH series of length 500.
GARCH <- sim_garch(inno,
                   model = list(mu = 0.000002,
                                omega = 0.00005,
                                alpha = 0.098839,
                                beta = 0.899506,
                                ar = 0.063666,
                                ma = NULL,
                                gamma = 0.12194,
                                delta = 1.85),
                   simple = FALSE)

 ### Visualising the resulting series
 GARCH \%>\% mutate(period = 1:n()) \%>\%
       gather(key, value, -period) \%>\%
       ggplot() +
       geom_line(aes(x = period, y = value, color = key)) +
       facet_wrap(~key, scales = "free_y", ncol = 1)

}
}
