% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_inno.R
\name{sim_inno}
\alias{sim_inno}
\title{sim_inno}
\usage{
sim_inno(
  corr,
  k = 252,
  mv_dist = "t",
  mv_df = 4,
  left_cop_weight = 0,
  left_cop_param = 5,
  marginal_dist = "norm",
  marginal_dist_model = NULL
)
}
\arguments{
\item{corr}{a correlation matrix defining the multivariate distribution of the simulated data. Note
that the number of variables simulated is equal to the number of columns in the correlation matrix
and increasing the left_cop_weight argument will weaken the degree to which the data resembles this
correlation structure.}

\item{k}{a positive integer indicating the number of time periods to simulate. Note that
the number of periods generated is actually equal to k + 5 as these extra observations may later
be needed when applying time series properties to the data, via the sim_garch function.}

\item{mv_dist}{a string specifying the multivariate distribution. Can be one of c("norm", "t")
referring to the multivariate normal and t distributions respectively. Default is "t".}

\item{mv_df}{degrees of freedom of the multivariate distribution, only influential when mv_dist = "t".
Default is 4.}

\item{left_cop_weight}{a positive value between zero and one indicating the weight applied to the
Clayton copula when creating the multivariate distribution. Note that a value between zero and one
essentially generates a hybrid distribution between the chosen mv_dist and the Clayton copula.
Therefore, the greater the left_cop_weight the less the data will reflect the correlation structure.
Default is set to 0.}

\item{left_cop_param}{a positive value indicating the parameter of the Clayton copula. Default is 4.}

\item{marginal_dist}{a string variable specifying the univariate distribution of each variable. Can
be one of c("unif", "norm", "t", "sgt") referring to the uniform(0,1), normal, student-t and
skewed-generalized-t distributions respectively. Default is "norm".}

\item{marginal_dist_model}{list containing the relevant parameters for the chosen marginal_dist. For
marginal_dist = "unif" marginal_dist_model doesn't effect output. marginal_dist = "norm" accepts a mean
and standard deviation with defaults list(mu = 0, sigma = 1) respectively. marginal_dist = "t" accepts
the non-centrality and degrees of freedom arguments, default values are list(mu = 0, df = 5).
marginal_dist = "sgt" accepts the mean, sd, lambda, p and q parameters
list(mu = 0, sigma = 1, lambda, p, q). Note lambda, p and q have no defaults and must therefore be set
by the user.}
}
\value{
A tibble with k+5 rows and columns containing the simulated values for each variable.
}
\description{
This function simulates a multivariate data set with a given correlation
structure. The multivariate distribution can be a weighted combinations between the normal
or student t and Clayton copula's. In creasing the weight on the Clayton copula (via the
left_cop_weight argument) will induce left tail dependence to the multivariate distribution.
The simulate univariate distributions can set to uniform, normal, student t or skewed generalized t
distributions. Note the function name "sim_inno", meaning simulate innovations, stems from the
i.i.d nature of the resulting series. Use the sim_asset_market function if you require
the data to exhibit time-series persistence.
}
\examples{

\dontrun{
### Generate ad hoc correlation matrix
cor <- gen_corr(D = 20, Clusters = "none")

### Simulating innovations
data <- sim_inno(cor,
                 k = 252,
                 mv_dist = "t",
                 mv_df = 5,
                 marginal_dist = "sgt",
                 marginal_dist_model = list(mu = 0,
                                            sigma = 1,
                                            lambda = -0.1,
                                            p = 2,
                                            q = Inf)
                 )
print(data)

}
}
