% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_market_with_progress.R
\name{sim_market_with_progress}
\alias{sim_market_with_progress}
\title{sim_market_with_progress}
\usage{
sim_market_with_progress(
  corr,
  k = 252,
  mv_dist = "t",
  mv_df = 3,
  left_cop_weight = 0,
  left_cop_param = 4,
  marginal_dist = "norm",
  marginal_dist_model = NULL,
  ts_model = list()
)
}
\value{
a tidy tibble containing a date, Asset and Return column.
}
\description{
This is an alternative version of the sim_market function that
includes functionality to include a progress bar when used in purrr::map functions
}
\note{
See ??sim_market for details on use.
}
\examples{

\dontrun{

library(MCmarket)
library(tidyverse)

### creating a correlation matrix to use as input in sim_asset_market
corr <- gen_corr(N = 20, Clusters = "none")


N <- 100
pb <- dplyr::progress_estimated(N)   # this must be named pb
market <-
      map_dfr(1:N,
              ~sim_market_with_progress(corr),
              .id = "Universe")

}
}
